refactor(cf): relocate cf org/space/route signal configs out of list framework
