+++ src/frontend/packages/cloud-foundry/src/features/cf/tabs/cf-organizations/cf-organization-summary/cloud-foundry-organization-summary.component.spec.ts
@@ -19,7 +19,7 @@ import { generateCFEntities } from '../../../../../cf-entity-generator';
 import { ActiveRouteCfOrgSpace } from '../../../cf-page.types';
 import { CloudFoundryOrganizationService } from '../../../services/cloud-foundry-organization.service';
 import { CloudFoundryEndpointService } from '../../../services/cloud-foundry-endpoint.service';
-import { CfOrgsSignalConfigService } from '../../../../../shared/components/list/list-types/org/cf-orgs-signal-config.service';
+import { CfOrgsSignalConfigService } from '../../../../../shared/signal-list-configs/org/cf-orgs-signal-config.service';
 import { CloudFoundryOrganizationSummaryComponent } from './cloud-foundry-organization-summary.component';
 
 describe('CloudFoundryOrganizationSummaryComponent', () => {
