+++ src/frontend/packages/cloud-foundry/src/shared/signal-list-configs/org/cf-orgs-signal-config.service.ts
@@ -2,11 +2,11 @@ import { DestroyRef, Injectable, Injector, Signal, WritableSignal, computed, eff
 import { HttpClient } from '@angular/common/http';
 import { firstValueFrom } from 'rxjs';
 import { ListStateStore } from '@stratosui/core';
-import { EndpointDataRegistry } from '../../../../../services/endpoint-data/endpoint-data.registry';
-import type { EndpointDataService } from '../../../../../services/endpoint-data/endpoint-data.service';
-import { ViewPipeline, SortSpec } from '../../../../../services/data-sources/view-pipeline';
-import { CnsiOrgsSource } from '../../../../../services/data-sources/cnsi-orgs-source';
-import type { StOrg } from '../../../../../services/endpoint-data/stratos-types';
+import { EndpointDataRegistry } from '../../../services/endpoint-data/endpoint-data.registry';
+import type { EndpointDataService } from '../../../services/endpoint-data/endpoint-data.service';
+import { ViewPipeline, SortSpec } from '../../../services/data-sources/view-pipeline';
+import { CnsiOrgsSource } from '../../../services/data-sources/cnsi-orgs-source';
+import type { StOrg } from '../../../services/endpoint-data/stratos-types';
 
 // Orgs list config service — single-CNSI analog to CfAppsSignalConfigService.
 // Unlike apps, the orgs view always lives under an explicit /cloud-foundry/:cnsi
